# Excalidraw perf